Added TinyMCE Editor, added dropdown in meni, home page set to display via model
