+++ src/main/webapp/WEB-INF/tags/pagination.tag
@@ -9,9 +9,10 @@
 <c:set var="startPage" value="${block * size + 1}" />
 <c:set var="endPage" value="${(block + 1) * size}" />
 <c:set var="endPage" value="${endPage > page.totalPages ? page.totalPages : endPage}" />
+<c:if test="${page.totalPages !=1}">
 <div class="pagination">
 	<c:if test="${block != 0}">
-		<a href="${url}?b=${block-1}&p=${(block-1)*size + 1}" style="margin-right:3px;">&lt;&lt;</a>
+		<a href="${url}?b=${block-1}&p=${(block-1)*size + 1}" style="margin-right:4px;">&lt;&lt;</a>
 	</c:if>
 	<c:forEach var="pageNumber" begin="${startPage}" end="${endPage}">
 		<c:choose>
@@ -27,7 +28,8 @@
 		</c:if>
 	</c:forEach>
 	<c:if test="${endPage != page.totalPages}">
-		<a href="?b=${block+1}&p=${(block+1)*size + 1}" style="margin-left:3px;">&gt;&gt;</a>
+		<a href="?b=${block+1}&p=${(block+1)*size + 1}" style="margin-left:4px;">&gt;&gt;</a>
 	</c:if>
 </div>
+</c:if>
