+++ src/main/webapp/WEB-INF/tiles/addStatus.jsp
@@ -33,3 +33,12 @@
 		</div>
 	</div>
 </div>
+
+ <script src="https://cdn.tiny.cloud/1/no-api-key/tinymce/5/tinymce.min.js" referrerpolicy="origin"></script>
+
+    <script>
+      tinymce.init({
+        selector: 'textarea',
+        pluging: "link"
+      });
+    </script>
