+++ src/main/webapp/WEB-INF/tiles/home.jsp
@@ -1,6 +1,8 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8"
     pageEncoding="ISO-8859-1"%>
 
-<div class="container">
-<h1>HOMEEEEE</h1>
+<div class="row">
+	<div class="col-md-6 col-md-offset-3 col-sm-8 col-sm-offset-2">
+	   ${statusUpdate.text}
+	</div>
 </div>
